+++ app/Mail/PasswordResetEmail.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace App\Mail;
+
+use App\Models\User;
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Mail\Mailable;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Support\Facades\Log;
+
+class PasswordResetEmail extends Mailable
+{
+    use Queueable, SerializesModels;
+
+    /**
+     * Create a new message instance.
+     *
+     * @return void
+     */
+    public function __construct(User $user)
+    {
+        $this->user = $user;
+    }
+
+    /**
+     * Build the message.
+     *
+     * @return $this
+     */
+    public function build()
+    {
+      Log::info('Sending Email');
+        return $this
+          ->subject('Password Reset Successful')
+          ->from(env('MAIL_DEFAULT_FROM'))
+          ->view('emails.account.passwordreset')
+          ->with([
+            'name' => $this->user['firstname'],
+          ]);
+    }
+}